.config/benchmarks/requirements.txt
psutil==5.9.5
py-cpuinfo==9.0.0
